handle sigterm gracefully
Previously, galera-init used the default behavior and terminated
immediately on SIGTERM. This led to an unclean exit of mysqld when
running in a container as the container would be torn down once the
controlling process (galera-init) exited.

Now a signal handler watches for SIGTERM and will cancel a Go Context
value. This context is managed by the start manager which will intiate a
graceful shutdown of the mysqld process once it detects a cancellation
event.

- Added KillCommand method to OsHelper to assist in signalling child
  processes.

[#161428221]

Signed-off-by: Caroline Taymor <ctaymor@pivotal.io>