reap mysqld children spawned for mysql_upgrade
Previously mysqld was run in a daemonize mode which left a zombie mysqld
process around after an upgrade completed successfully, and galera-init
started the "normal" long-lived mysqld.

Now the Upgrader does not daemonize mysqld but rather keeps an active
process handle, upgrades mysql and then waits for the shutdown process
of the "upgrade" mysqld to complete. This ensures a confusing second
"zombie" mysqld does not persist in a long-lived bpm container.

[#161428221]

Co-authored-by: Caroline Taymor <ctaymor@pivotal.io>