WIP: support deferred plugin configuration
When initializing a MySQL data directory the first time by running the
mysqld --initialize command, mysqld disables loading plugins.  This has
the effect of also making any plugin configuration options in the my.cnf
invalid / "unknown" when the plugin is not loaded.  This will cause
mysqld --initialize to fail.

This happens in pxc-release when audit logging is enabled for a fresh
deploy.  mysqld --initialize starts with the audit plugin disabled, but
then fails with an error like:

  > [ERROR] [MY-000067] [Server] unknown variable 'audit-log-file=foo'.

To support plugin configuration in the my.cnf, when starting mysqld
normally uses --defaults-group-suffix=_plugin.   This allows a my.cnf to
specify a section like:

[mysqld_plugin]
plugin-load-add = audit_log.so
audit-log-file  = foo

And this is not parsed by mysqld --initialize and avoids these errors.

WIP because:

- Think about the suffix name.  Maybe a more generic name like "runtime"
  or "deferred" makes more sense.  Although this _probably_ only applies
  to plugins.